pyproject.toml
[project]
name = "trainformer"
version = "0.1.0"
description = "Python-first training library for deep learning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core
    "torch>=2.0",
    "accelerate>=0.25",

    # Vision
    "timm>=0.9",
    "torchvision>=0.15",

    # NLP
    "transformers>=4.36",
    "peft>=0.7",
    "datasets>=2.14",

    # Evaluation
    "faiss-cpu>=1.7",

    # Logging
    "wandb>=0.16",

    # CLI
    "fire>=0.5",

    # Utilities
    "cloudpickle>=3.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "ruff>=0.1"]
qlora = ["bitsandbytes>=0.41"]
tensorboard = ["tensorboard>=2.14"]
mlflow = ["mlflow>=2.8"]
all = ["trainformer[qlora,tensorboard,mlflow]"]

[project.scripts]
trainformer = "trainformer.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
